+++ src/components/listItem/index.js
@@ -84,6 +84,8 @@ ListItem.propTypes = {
     hover: PropTypes.bool,
     noDivider: PropTypes.bool,
     checkbox: PropTypes.bool,
+    item: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
+    onClick: PropTypes.func,
     subTitle: PropTypes.string,
     icon: PropTypes.string,
     tooltip: PropTypes.string,